Improvements to reduce number of files
