+++ 6-http-server-module/about.html
@@ -6,6 +6,6 @@
 	<link rel="stylesheet" href="style.css">
 </head>
 <body class="about">
-	<h1>About h1age</h1>
+	<h1>About Page</h1>
 </body>
 </html>
